chore: Move canonical facts to own internal package
* Card ID: CCT-1591
* Moved canonical_facts.go and canonical_facts_test.go
  to ./internal/canonical_facts/
* Modified usage of package accordingly
* No functional changes
